The basics of async

+++ python_async_function/0-basic_async_syntax.py
@@ -12,17 +12,17 @@ import asyncio
 
 
 # Couroutine function
-async def wait_random(max_delay: float = 10) -> float:
+async def wait_random(max_delay: int = 10) -> float:
     """
     Asynchronous coroutine that waits for a random delay
     between 0 and max_delay
     (included and float value) seconds and eventually returns it.
 
     Args:
-        max_delay (float, optional): The maximum delay. Defaults to 10.
+        max_delay (int, optional): The maximum delay. Defaults to 10.
 
     Returns:
-        float: The actual delay used.
+        int: The actual delay used.
     """
     delay = random.uniform(0, max_delay)
     await asyncio.sleep(delay)
